Work around malloc related crashes on ARM
We're seeing malloc heap corruption while running the interpreter under ARM. In
addition, we're seeing segfaults within jemalloc while running gen-class-map.

Linking in "mcheck" (which enables malloc debugging) seems to make the problem
go away mysteriously, so we'll just leave that in for now. Disable jemalloc for
ARM builds for now as well.