Fix instruction smashing on ARM
Fix a few bugs in instruction smashing on ARM. Namely, when smashing a JMP
target, you don't want to take into account the full size of the block, because
its always computed from a return address (which sits in the middle of the
block). Also, the beginning of the block can be padded with NOPs, so don't
assume in the ARM vasm emitter that the frontier is the beginning of the
block. Instead use the return value from the code which emits the smashable
jumps.