Hook up vasm enterTCHelper for ARM
Rather than have the simulator set up first to do the equivalent of
enterTCHelper, simply set up the initial simulator state as if enterTCHelper had
just been called, and let the simulator run on the native enterTCHelper for
ARM. This reduces the differences between native ARM and simulated ARM. Add
support for calling enterTCHelper natively (without the simulator).

Also, have X64 use the enterTCHelper unique-stub machinery rather than calling
the assembly code directly. For now, the unique-stub is the assembly, but this
makes it easier to vasm-fy it in the future.